+++ docs/source/api_references/commands.rst
@@ -1,0 +1,6 @@
+===============================
+Command Utilities API Reference
+===============================
+
+.. automodule:: toolbox.commands
+   :members:
