+++ docs/source/api_references/embeds.rst
@@ -1,0 +1,6 @@
+=============================
+Embed Utilities API Reference
+=============================
+
+.. automodule:: toolbox.embeds
+   :members:
