+++ docs/source/api_references/messages.rst
@@ -1,0 +1,6 @@
+===============================
+Message Utilities API Reference
+===============================
+
+.. automodule:: toolbox.messages
+   :members:
