+++ docs/source/api_references/strings.rst
@@ -1,0 +1,6 @@
+==============================
+String Utilities API Reference
+==============================
+
+.. automodule:: toolbox.strings
+   :members:
