--- docs/source/api_references/utils.rst
@@ -1,6 +1,0 @@
-=======================
-Utilities API Reference
-=======================
-
-.. automodule:: toolbox.utils
-   :members:
